+++ examples/lookback_and_retries/main.py
@@ -21,7 +21,7 @@ from execute import execute
 @load_models
 def main(models: list[Model], debug: bool) -> None:
     for model in models:
-        intervals = model.infer_intervals()
+        intervals = model.intervals
 
         print(f"\n{model.target.full_name}")
         print(f"  interval_expression : {model.batching.interval.expression}")